LayoutTests/imported/w3c/web-platform-tests/cors/preflight-cache-partitioning.sub.window-expected.txt


Harness Error (FAIL), message = Error: assert_equals: The CORS preflight wasn't sent expected "pass" but got "fail"

TIMEOUT The preflight cache should be partitioned Test timed out

Harness Error (FAIL), message = Error: assert_equals: The CORS preflight wasn't sent expected "pass" but got "fail"

TIMEOUT The preflight cache should be partitioned Test timed out
